refactor!(nova): Move components out of 'components' folder in /challenges
