+++ apps/nova/src/app/[locale]/(dashboard)/challenges/Countdown.tsx
