+++ apps/nova/src/app/[locale]/(dashboard)/challenges/DurationDisplay.tsx
