+++ apps/nova/src/app/[locale]/(dashboard)/challenges/TimeProgress.tsx
